beam_validator: Remove the file/1 and files/1 functions
Before the beam_validator was added as compiler pass, it was a
standalone module that could analyse existing .beam files and .S
files.

Even though beam_validator has been part of the compiler for many
releases, it still supports the analysis of .beam and .S files.

To reduce the code bloat and to improve coverage of beam_validator,
remove the file/1 and files/1 functions and all associated help
functions. We'll need to update the test suite, since some of the
checked in .S files have errors that beam_validator ignores, but
that will not be accepted when running them throught the compiler
using the 'from_asm' option. In particular, we will need to export
all functions that should be validated (since the beam_clean pass
will remove any function that is not possible to call).