v3_life: Combine literal/2 and literal2/2
For unclear reasons, there are two functions in v3_life that are
almost identical: literal/2 and literal2/2. literal/2 is used
for expressions and literal2/2 for patterns.

It turns out that literal2/2 can do everything that literal/2 can
do, except that it transforms maps differently.

If we adjust v3_codegen to accept the same format of maps in
expressions and patterns, we can rename literal2/2 to literal/2
and use it for expressions and patterns.