beam_type: Use the complete register map when calculating liveness
When calculating the number of live registers for allocation
instruction, it is not always safe to start with the number of live
registers at the start of the block. We will need to use the register
map to know whether there are any holes (dead registers) that are not
subsequently filled.

If the allocation instruction already has a number of live registers
calculated, there is nothing to be gained by raising it.