v3_codegen: Teach the put_map_* instructions to reuse source registers
The put_map_assoc and put_map_exact instructions in the run-time
system will support that the target register is the same as one of
the source registers. Teach the code generator to take advantage
of that.

The disadvantages of not reusing register when possible is that the
garbage collector may retain dead terms longer than necessary.