PER, UPER: Eliminate compiler warning for SeqPrim.asn1
The BEAM compiler now warns more aggressively for expressions whose
values are not used. For example, the generated Erlang code for
the following ASN.1 code will now cause a warning:

  Seq ::= SEQUENCE {
    ...,
    os OCTET STRING (SIZE (17000..30000))
  }

The generated Erlang code looks similar to the following:

  Enc9@bin = iolist_to_binary(Enc9@output),
  Enc9@len = byte_size(Enc9@bin),
  [align|encode_fragmented(Enc9@bin, 8)]

The variable Enc9@len is not used and the BEAM compiler will
complain that the value returned from the byte_size/1 call is
not used.

Improve the optimization of the intermediate code to eliminate
the assignment to the unused variable.