binary_module_SUITE: Remove unnecessary calls to the binref module
The compile_pattern/1 function in the binref module is a dummy
function. Therefore, calling binref:match/2 once with the plain
pattern and once with a "pre-compiled" pattern will take twice as
long, but will not help us find more errors.

I shaved off one fourth of the running time for random_ref_comp/1
by eliminating the unnecessary calls to the binref module.