Optimize unicode_SUITE:ex_binaries_errors* test cases
Refactor the code so that the error test cases for utf16 can
share the code for big and little endian. Do the same for utf32.

The major optimization is the observation that the following code:

  byte_size(unicode:characters_to_binary(List, unicode, {utf16,big})

can be replaced with:

  2*length(List)

which is much faster. (That optimization is not safe in general,
but for the lists used in the error tests cases, it is.)