Teach the loader to pre-compute the hash value for single-key lookups
Let the loader pre-compute the hash value when a single, literal key
is matched as in:

  #{<<"some_key">>:=V} = Map

In my measurements, this optimization resulted in a 30 percent
speedup for short binary keys.

Unfortunately, this optimizization makes no difference for small
maps with less than 32 keys, since the hash value is not used.
Still, there are the following use cases:

* A map used instead of a record with more than 32 entries. I have
seen some applications with huge records.

* Lookup in JSON dictionaries represented as maps.
The hash value will only be used when the map is a hash map
(currently, that means at least 32 entries).