Raise more descriptive error messages for failed map operations
According to EEP-43 for maps, a 'badmap' exception should be
generated when an attempt is made to update non-map term such as:

   <<>>#{a=>42}

That was not implemented in the OTP 17.

José Valim suggested that we should take the opportunity to
improve the errors coming from map operations:

  http://erlang.org/pipermail/erlang-questions/2015-February/083588.html

This commit implement better errors from map operations similar
to his suggestion.

When a map update operation (Map#{...}) or a BIF that expects a map
is given a non-map term, the exception will be:

  {badmap,Term}

This kind of exception is similar to the {badfun,Term} exception
from operations that expect a fun.

When a map operation requires a key that is not present in a map,
the following exception will be raised:

  {badkey,Key}

José Valim suggested that the exception should be
{badkey,Key,Map}. We decided not to do that because the map
could potentially be huge and cause problems if the error
propagated through links to other processes.

For BIFs, it could be argued that the exceptions could be simply
'badmap' and 'badkey', because the bad map and bad key can be found in
the argument list for the BIF in the stack backtrace. However, for the
map update operation (Map#{...}), the bad map or bad key will not be
included in the stack backtrace, so that information must be included
in the exception reason itself. For consistency, the BIFs should raise
the same exceptions as update operation.

If more than one key is missing, it is undefined which of
keys that will be reported in the {badkey,Key} exception.