erts: Ensure hashing of zero is consistent
Erlang treats positive and negative zero as equal, meaning,

    true = 0.0 =:= 0.0/-1

However, Erlangs hash functions: hash, phash and phash2 did not
reflect this behaviour. Meaning, the hash values produced by
the different hash functions would not be identical for positive
and negative zero.

This commit ensures that hash value of positive zero is always
produced regardless of the signedness of the zero float, i.e.

    true = erlang:phash2(0.0) =:= erlang:phash2(0.0/-1)