ssh: fix loop bug for zero maximum_packet_size.
A channel could be opened with the 'max packet size' set to 0.  The behaviour is not defined in the ssh rfc:s, and openssh does not handle that case in some special way besides looping.

The introduced solution just "keep quiet" but removes the eternal loop that caused Erlang to core dump.