Fix gen_tcp:shutdown/2 by making it asynchronous
If the driver queue is empty, or the user is requesting a 'read'
shutdown, then the shutdown() syscall is performed synchronously, as
per the old version of shutdown/2.

However, if the user is requesting a 'write' or 'read_write' shutdown,
and there is data in the driver queue for the socket, then the
shutdown() syscall is delayed and handled asynchronously when the
driver queue is written out.

This version of shutdown solves a number of issues with the old
version. The two main solutions it offers are:

 * It doesn't block when the TCP peer is idle or slow. This is the
   expected behaviour when shutdown() is called: the caller needs
   to be able to continue reading from the socket, not be prevented
   from doing so.

 * It doesn't truncate the output. The current version of
   gen_tcp:shutdown/2 will truncate any outbound data in the driver
   queue after about 10 seconds if the TCP peer is idle of slow. Worse
   yet, it doesn't even inform anyone that the data has been
   truncated: 'ok' is returned to the caller; and a FIN rather than
   an RST is sent to the TCP peer.

For a detailed description of all the problems with the old version
of shutdown, please see the EEP Light that was written to justify
this patch.