beam_type: Eliminate redundant calls to checkerror_1/2
Profiling shows that the excution time for checkerror_1/2 could
be be near the top even for modules without any floating point
operations.

It turns out that the complexity of simplify_float_1/4 is quadratic.
checkerror/1 is called with the growing accumulator for each
iteration. checkerror/1 will traverse the entire accumulated list
*unless* some floating point operations are used.

We can avoid this situation if we only call checkerror/1 when there
are live floating point registers. We can also avoid calling flush/3
if there are no live floating point registers.