zip: Suppress a dialyzer warning for an unmatched return
A warning for an unmatched return was introduced in aaa7c917.

The caller of openzip_close/1 ignores the return value, which
could be an error tuple. In this situation, handling a fatal
error, it is OK to ignore the return value; thus, we only need
to make it explicit that we are ignoring the return value.