Add p(Term) shell command for printing terms
The shell does two things by default: it formats terms
resulting from expressions so that they take less than
80 columns when possible; and it restricts the printed
term to a specific depth and length.

There are two main alternatives to printing terms. The
shell rp/1 function and the io:format/* functions will
print the full term but will still do potentially very
intensive formatting.

The erlang:display/1 function does what we want except
it will not work on a remote shell as it prints on the
remote node rather than the local one.

This function settles on erlang:display/1 for printing
as there are no solution that works in all cases.

This function will also work when the io is taken over
by eunit or common_test and can therefore be used in a
shell session while they are running.