erts,hipe: Fix bug in binary matching of writable binary
Seen symptom: Hipe compiled code with <<C/utf8, ...>> = Bin does sometimes
not match even though Bin contains a valid utf8 character. There might be
other possible binary matching symptoms, as the problem is not utf8
specific.

Problem: A writable binary was not "emasculated" when the matching started
(as it should) by the hipe compiled code.

Fix: Add a new primop emasculate_binary(Bin) that is called when
a matchstate is created.

ToDo: There are probably room for optimization. For example only call
emasculate_binary if ProcBin.flags is set.