beam_utils: Be less conservative about liveness for exit instructions
beam_utils used to be overly conservative about liveness for
exit instructions such as:

  call_ext erlang:exit/1

beam_utils would consider all y registers to be used, to avoid
overwriting a catch or try tag. That does not seem to be a real
risk.

However, we miss opportunities for stack trimming if we consider
y registers used by an exit instruction.