dialyzer: Modify the handling of parametrized opaque types
In OTP 17 it is possible to mix types such as dict:dict() and
dict:dict(_, _) outside of the dict module (and similarly for some
other opaque types in STDLIB), but the results are unfortunately
possibly invalid warnings in users' code. In OTP 18 parameterized
opaque types with the same name but with different number of
parameters are no longer compatible when seen from outside of the
module where the types are declared.

The types in STDLIB have been updated accordingly; for instance
  -opaque dict() :: dict(_, _).
has been replaced by
  -type dict() :: dict(_, _).