beam_asm: Speed up encoding of large numbers
The misc_SUITE:integer_encoding/1 test case is annoyingly slow.

Rewrite the encoding of integers in beam_asm to use the
binary:encode_unsigned/1 BIF.

Also tweak the test case itself. Scale the down the maximum
size of the numbers being generated, but also add test of
numbers around boundaries of power of two (which are the numbers
most likely to expose bugs in the encoding).