Add dialyzer_plt:relocate/3
dialyzer_plt:relocate(OldFile, NewFile, BaseDirs) will change the file
paths in given plt file OldFile and save it to NewFile.  The file paths
in NewFile will all be found under the directories in BaseDirs.

This is useful for erlang libraries.  A plt file for all beam files can
be included in the library distribution.  Users of the library can then
relocate the plt file according to where ever they extracted the
library.  This is much faster than building a plt from the beams.