kernel application: Start code server first
When the code server has started, it will read the .beam files
for all loaded modules to see whether they contain native code
that should be loaded. To make that as fast as possible when
we have no native code, start the code server as early as possible
to minimize the number .beam files that must be read from disk.

The code server does not use the 'file' module, so it is OK to
start the code server before starting the file_server process.

On my computer, the time for checking the loaded modules for
native code is reduced from roughly 2 ms to 1.5 ms.