common_test: Recurse when there is more data in netconf
When several packets where receive in one packet ct_netconf
failed to deliver them to the user.

For example several subscritiption message could be in the buffer
but only the first was sent to the user.

Error handling could be improved, maybe the connection should
be closed when unparseable packet arrives or timeout occurs.