Don't confuse Result-Code and Experimental-Result
Decode of an answer message not setting the E-bit, and containing
Experiment-Result but not Result-Code, identified Result-Code as the
erroneous when Erroneous-Result-Code was 3xxx. Here's an example (from
trace) of a the errors field after decode:

  [{5004,
    {diameter_avp,undefined,undefined,false,false,undefined,'Result-Code',
     3001,undefined,undefined}}],

The diameter_avp was just constructed from the AVP name and decoded
result, without regard for which result code AVP contained the value.
Fix by extracting the AVP from the incoming message.