ssl: Refactored and corrected sni_fun handling
Dialyzer warned about the incorrect match of Packets. Code
was refactored and the problem avoided in the process.

Dialyzer warned that the empty tuple is not a function as the contract
said it should be. Changed the handling of the sni_fun default value to be
undefined and added it to the contract.