erts: Disable float exceptions for clang/llvm
Change erts/configure.in to force-disable FP exceptions if
the VM is compiled by clang/llvm.  clang/llvm generates FP
code which does not work with FP exceptions (whether unmasked
as used in the Erlang VM, or masked followed by tests of
which are signalled).  This is a known long-standing problem.