sys_core_fold: Eliminate warnings for unused terms
The optimization introduced in 0a0d39d351fc would cause spurious
warnings of the type: "a term is constructed, but never used".

To avoid the warning, we must mark not only tuples and lists as
compiler_generated, but also each element. We must also propagate
compiler_generated annotations in lets. For example, if we have:

  let <X -| ['compiler_generated']> = 42 in X + 1

we must propagate the compiler_generated annotation to the literal
when do constant propagation:

  42 -| ['compiler_generated'] + 1