ssh: option for handling the SSH_MSG_DEBUG message's printouts
A fun could be given in the options that will be called whenever
the SSH_MSG_DEBUG message arrives.  This enables the user to
format the printout or just discard it.

The default is changed to not print the message. In RFC4253
printing is a SHOULD, but our new default is to protect logs
from dos attacs.